+++ src/common/base/src/runtime/mod.rs
@@ -28,7 +28,9 @@ pub use catch_unwind::catch_unwind;
 pub use catch_unwind::CatchUnwindFuture;
 pub use global_runtime::GlobalIORuntime;
 pub use global_runtime::GlobalQueryRuntime;
+pub use memory::set_alloc_error_hook;
 pub use memory::MemStat;
+pub use memory::GLOBAL_MEM_STAT;
 pub use runtime::block_on;
 pub use runtime::execute_futures_in_parallel;
 pub use runtime::match_join_handle;
@@ -41,11 +43,9 @@ pub use runtime::Dropper;
 pub use runtime::Runtime;
 pub use runtime::TrySpawn;
 pub use runtime::GLOBAL_TASK;
-pub use runtime_tracker::set_alloc_error_hook;
 pub use runtime_tracker::LimitMemGuard;
 pub use runtime_tracker::ThreadTracker;
 pub use runtime_tracker::TrackedFuture;
 pub use runtime_tracker::UnlimitedFuture;
-pub use runtime_tracker::GLOBAL_MEM_STAT;
 pub use thread::Thread;
 pub use thread::ThreadJoinHandle;
